Group Hero tests in same file
