+++ apps/heroes_server/test/game/hero_test.exs
@@ -1,7 +1,7 @@
 defmodule Game.HeroTest do
   use ExUnit.Case, async: true
 
-  alias Game.Hero
+  alias Game.{Hero, HeroServer}
   alias GameError.BadCommand
 
   @board_4x4 GameBoards.Test4x4
@@ -159,7 +159,7 @@ defmodule Game.HeroTest do
     tile = Map.get(context, :tile, {1, 1})
 
     opts = [board: board, tile: tile]
-    [hero: start_supervised!({Hero, opts})] ++ opts
+    [hero: start_supervised!({HeroServer, opts})] ++ opts
   end
 
   defp control(pid, commands) do
@@ -169,5 +169,42 @@ defmodule Game.HeroTest do
     end
 
     Enum.reduce(commands, :acc, fn cmd, _ -> unwrap.(cmd) end)
+  end
+end
+
+defmodule Game.HeroTestSync do
+  use ExUnit.Case
+
+  alias Game.{Hero, HeroServer, HeroSupervisor}
+
+  describe "A player controlling a hero" do
+    setup context do
+      for tile <- context.tiles, into: %{} do
+        hero = create_hero(GameBoards.Test4x4, tile)
+        {tile, hero}
+      end
+    end
+
+    @tag tiles: [{1, 0}, {1, 1}, {2, 2}, {3, 0}]
+    test "can attack all enemies at once", %{{1, 1} => hero} = context do
+      assert {:ok, :launched} = Hero.control(hero, :attack)
+      :timer.sleep(50)
+      assert {:ok, {0, 1}} = Hero.control(hero, :left)
+
+      dead_enemy = Map.get(context, {1, 0})
+      assert {:error, :noop} = Hero.control(dead_enemy, :up)
+
+      dead_enemy = Map.get(context, {2, 2})
+      assert {:error, :noop} = Hero.control(dead_enemy, :left)
+
+      live_enemy = Map.get(context, {3, 0})
+      assert {:ok, {3, 1}} = Hero.control(live_enemy, :up)
+    end
+  end
+
+  defp create_hero(board, tile) do
+    child = {HeroServer, [board: board, tile: tile]}
+    {:ok, pid} = DynamicSupervisor.start_child(HeroSupervisor, child)
+    pid
   end
 end
